AB# Fix Dev environment issues
