+++ UI/apps/public/public/img/icons/site.webmanifest
@@ -3,12 +3,12 @@
   "short_name": "",
   "icons": [
     {
-      "src": "/android-chrome-192x192.png",
+      "src": "/img/icons/android-chrome-192x192.png",
       "sizes": "192x192",
       "type": "image/png"
     },
     {
-      "src": "/android-chrome-512x512.png",
+      "src": "/img/icons/android-chrome-512x512.png",
       "sizes": "512x512",
       "type": "image/png"
     }
