+++ UI/libs/core-admin/src/lib/components/navigation/SideBar.vue
@@ -130,29 +130,19 @@
 import LoginButton from '@core-admin/components/login/LoginButton.vue';
 import Routes from '@core-admin/router/routes';
 import SearchBar from '@core-admin/components/search/SearchBar.vue';
-import { useAppConfigStore } from '@shared-ui/stores/appConfig';
+import { ref } from 'vue';
 import { useAppointmentsStore } from '@shared-ui/stores/appointmentsStore';
 import { useBrandStore } from '@core-admin/stores/brandStore';
+import useEnvName from '@shared-ui/composables/useEnvName';
 import { useQuery } from '@tanstack/vue-query';
-import { computed, ref } from 'vue';
 
 const drawer = ref(true);
 const aptStore = useAppointmentsStore();
-const appStore = useAppConfigStore();
 const brandStore = useBrandStore();
 
 const { isLoading } = useQuery(['logo']);
 
-const getAppTitle = computed(() => {
-  switch (appStore.getAppConfig.environmentName) {
-    case 'DEV':
-      return 'CCW (DEV)';
-    case 'QA':
-      return 'CCW (QA)';
-    default:
-      return 'CCW';
-  }
-});
+const getAppTitle = useEnvName();
 </script>
 
 <style lang="scss" scoped>
