+++ UI/libs/core-public/src/lib/components/navbar/NavBar.vue
@@ -5,7 +5,11 @@
     class="flex-grow-0 white--text"
     clipped-right
   >
-    <ConnectionStatus class="ml-4" />
+    <router-link :to="{ path: Routes.HOME_ROUTE_PATH }">
+      <div class="text-h6 white--text">
+        {{ getAppTitle }}
+      </div>
+    </router-link>
     <ThemeMode class="ml-4" />
     <v-spacer></v-spacer>
     <LoginButton> </LoginButton>
@@ -13,9 +17,12 @@
 </template>
 
 <script setup lang="ts">
-import ConnectionStatus from '@shared-ui/components/status/ConnectionStatus.vue';
 import LoginButton from '@core-public//components/login/LoginButton.vue';
+import Routes from '@core-public/router/routes';
 import ThemeMode from '@shared-ui/components/mode/ThemeMode.vue';
+import useEnvName from '@shared-ui/composables/useEnvName';
+
+const getAppTitle = useEnvName();
 </script>
 
 <style lang="scss" scoped>
