+++ UI/libs/core-public/src/lib/router/routes.ts
@@ -36,8 +36,8 @@ export default class Routes {
     return `/qualifying-questions`;
   }
 
-  static get RECIEPT_PATH() {
-    return '/reciept';
+  static get RECEIPT_PATH() {
+    return '/receipt';
   }
   static get RENEW_APPLICATION_ROUTE_PATH() {
     return `/renew-application`;
